Add files via upload
#Employee Attrition: Exploratory Data Analysis (EDA)

# The dataset collected contains the following fields that describe employee attrition for a company:

# satisfaction_level: value between 0 - 1 that describes how satisfied the employee was at their current role
# last_evaluation: value between 0 - 1 that describes how well the employee performed
# number_project: number of projects the employee was on
# average_montly_hours: average number of hours worked monthly time_spend_company: number of years at the company
# work_accident: boolean for if the employee has been in an accident at work
# left: boolean that describes if the employee has left the company
# role: employee's role
# salary: low, medium, high --> indicates how high employee salary was
    
#Can we describe which factors are the strongest indicator if an employee is going to attrit?

Data Source: https://docs.google.com/spreadsheets/d/1lJL6IYuQqBV8xTdc5IZPjz_pKRluSUeHuKUOJss7uB8/edit#gid=1240005

Now that we have the Pearson's R and the p-values, we can stack rank the variables. Pearson's R tells us how linearly correlated two variables are, where as p-value tells use how significant our Pearson's R is. The p-value ranges from 0 - 1.0, with 0.00 being highly significant and 1.0 being not significant. The p-value takes into account both the strength of the correlation, R, as well as the number of samples. For example, if you had only two samples, you could easily fit a line through them, but your p-value would be large because 2 samples just aren't enough to tell you what's going on. Even if we had a nearly perfect correlation, we will not get statistical significance. In total we'll have four scenarios that could, below summarized the scenarios and their interpretations:

--Low R and low p-value (typically, p-value <= 0.05): means that we're not able to explain much of the variation of the data but it is significant (second best)

--Low R and high p-value (typically, p-value > 0.05): means that we're not able to explain a lot of variation and it's not significant (worst scenario)

--High R and low p-value: means that we're able to explain a lot of variation within the data and is significant (the best scenario)

--High R and high p-value: means that we're able to explain a lot of variation within the data but is not significant (second to worst scenario)

Now that we have the scenarios outlined, let's go through our results and stack rank each factor. In our scenario, a lot of our factors are not well-correlated with attrition, however we can still stack rank based on Pearson's R and p-value. To do this, first I stack ranked by p-value because a high p-value shows little significance regardless of what the Pearson's R value is. Next I ranked based on the absolute value for Pearson's R. 

We're doing this because if we had two variables (e.g. one with a 0.5 R value and the other with -0.5 R), they would have the same correlation strength, one is positive and the other is negative.